+++ test/SharpLink.LoadTest.Tests/GlobalUsings.cs
@@ -1,0 +1,7 @@
+global using System;
+global using System.Collections.Generic;
+global using System.Threading;
+global using System.Threading.Tasks;
+global using SharpLink.LoadTest;
+global using SharpLink.Runtime;
+global using TUnit.Core;